Move url string helpers to a file
