+++ src/sentry/notifications/utils/__init__.py
@@ -1,16 +1,13 @@
 from __future__ import annotations
 
 import logging
-import time
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, Iterable, Mapping, MutableMapping, Sequence, cast
 
 from django.db.models import Count
-from django.utils.http import urlencode
 from django.utils.safestring import mark_safe
 
 from sentry import integrations
-from sentry.incidents.models import AlertRuleTriggerAction
 from sentry.integrations import IntegrationFeatures, IntegrationProvider
 from sentry.models import (
     Activity,
@@ -32,7 +29,6 @@ from sentry.models import (
 from sentry.notifications.notify import notify
 from sentry.notifications.utils.participants import split_participants_and_context
 from sentry.utils.committers import get_serialized_event_file_committers
-from sentry.utils.http import absolute_uri
 
 if TYPE_CHECKING:
     from sentry.eventstore.models import Event
@@ -128,59 +124,6 @@ def summarize_issues(
 
         rv.append({"message": EventError(msg_d).message, "extra_info": extra_info})
     return rv
-
-
-def get_email_link_extra_params(
-    referrer: str = "alert_email",
-    environment: str | None = None,
-    rule_details: Sequence[NotificationRuleDetails] | None = None,
-    alert_timestamp: int | None = None,
-) -> dict[int, str]:
-    alert_timestamp_str = (
-        str(round(time.time() * 1000)) if not alert_timestamp else str(alert_timestamp)
-    )
-    return {
-        rule_detail.id: "?"
-        + str(
-            urlencode(
-                {
-                    "referrer": referrer,
-                    "alert_type": str(AlertRuleTriggerAction.Type.EMAIL.name).lower(),
-                    "alert_timestamp": alert_timestamp_str,
-                    "alert_rule_id": rule_detail.id,
-                    **dict([] if environment is None else [("environment", environment)]),
-                }
-            )
-        )
-        for rule_detail in (rule_details or [])
-    }
-
-
-def get_group_settings_link(
-    group: Group,
-    environment: str | None,
-    rule_details: Sequence[NotificationRuleDetails] | None = None,
-    alert_timestamp: int | None = None,
-) -> str:
-    alert_rule_id: int | None = rule_details[0].id if rule_details and rule_details[0].id else None
-    return str(
-        group.get_absolute_url()
-        + (
-            ""
-            if not alert_rule_id
-            else get_email_link_extra_params(
-                "alert_email", environment, rule_details, alert_timestamp
-            )[alert_rule_id]
-        )
-    )
-
-
-def get_integration_link(organization: Organization, integration_slug: str) -> str:
-    # Explicitly typing to satisfy mypy.
-    integration_link: str = absolute_uri(
-        f"/settings/{organization.slug}/integrations/{integration_slug}/?referrer=alert_email"
-    )
-    return integration_link
 
 
 @dataclass
